+++ freemarker-generator-base/src/main/java/org/apache/freemarker/generator/base/file/RecursiveFileSupplier.java
@@ -71,6 +71,7 @@ public class RecursiveFileSupplier implements Supplier<List<File>> {
         return sources.stream()
                 .map(this::resolve)
                 .flatMap(Collection::stream)
+                .sorted()
                 .collect(toList());
     }