+++ freemarker-generator-cli/src/test/java/org/apache/freemarker/generator/cli/ExamplesTest.java
@@ -114,9 +114,15 @@ public class ExamplesTest extends AbstractMainTest {
     }
 
     @Test
+    public void shouldTransformTemplateDirectory() throws IOException {
+        assertTrue(execute("-t site/template").contains("server.name=somehost"));
+        assertTrue(execute("-t site/template -PNGINX_HOSTNAME=localhost").contains("server.name=localhost"));
+    }
+
+    @Test
     @Ignore("Manual test to check memory consumption and resource handling")
     public void shouldCloseAllResources() throws IOException {
-        for (int i = 0; i < 5000; i++) {
+        for (int i = 0; i < 500; i++) {
             shouldRunInfo();
             shouldRunDemoExamples();
             shouldRunCsvExamples();
@@ -128,6 +134,7 @@ public class ExamplesTest extends AbstractMainTest {
             shouldRunXmlExamples();
             shouldRunGrokExamples();
             shouldRunInteractiveTemplateExamples();
+            shouldTransformTemplateDirectory();
             shouldRunWithExposedEnvironmentVariableExamples();
         }
     }