+++ freemarker-generator-cli/src/test/java/org/apache/freemarker/generator/cli/config/ConfigurationSupplierTest.java
@@ -59,6 +59,8 @@ public class ConfigurationSupplierTest {
     }
 
     private SettingsBuilder settingsBuilder() {
-        return Settings.builder().setTemplateName(ANY_TEMPLATE_NAME).setWriter(new StringWriter());
+        return Settings.builder()
+                .setTemplateNames(singletonList(ANY_TEMPLATE_NAME))
+                .setWriter(new StringWriter());
     }
 }